+++ docsig/_decorators.py
@@ -27,9 +27,9 @@ def parse_msgs(func: _WrappedFuncType) -> _WrappedFuncType:
     @_functools.wraps(func)
     def _wrapper(*args: str | _Path, **kwargs: _t.Any) -> str | int:
         disable = _E.from_codes(kwargs.get("disable", [])) or None
-        targets = _E.from_codes(kwargs.get("targets", [])) or None
+        target = _E.from_codes(kwargs.get("target", [])) or None
         kwargs["disable"] = disable
-        kwargs["targets"] = targets
+        kwargs["target"] = target
         return func(*args, **kwargs)
 
     return _wrapper
@@ -79,7 +79,7 @@ def validate_args(func: _FuncType) -> _WrappedFuncType:
                         f"unknown option to disable '{message.description}'"
                     )
 
-            for message in kwargs.get("targets") or []:
+            for message in kwargs.get("target") or []:
                 if not message.isknown:
                     stderr.append(
                         f"unknown option to target '{message.description}'"